% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_base.R
\name{assert_base}
\alias{assert_base}
\alias{chk_assert_func}
\alias{chk_insist_func}
\title{General data audit function to create other specific audit functions.}
\usage{
chk_assert_func(.data, .function, .variables)

chk_insist_func(.data, .function, .variables)
}
\arguments{
\item{.data}{Data to check for errors.}

\item{.function}{\code{assertr} function to apply to the \code{\link[assertr:assert]{assertr::assert()}} command.}

\item{.variables}{The columns to run the audit checks on.}
}
\value{
Adds \code{assertr_errors} to data.frame.
}
\description{
Thin wrapper around the \code{\link[assertr:assert_]{assertr::assert_()}} and \code{\link[assertr:insist_]{assertr::insist_()}} functions.
}
\examples{
library(magrittr)
library(assertr)

flaws <- iris \%>\%
chk_assert_func(within_bounds(5, 7), "Sepal.Length")
aud_report(flaws)

flaws <- iris \%>\%
chk_insist_func(within_n_mads(1), "Sepal.Length")
aud_report(flaws)

}
\seealso{
See documentation on either \code{\link[assertr:assert]{assertr::assert()}} or \code{\link[assertr:insist]{assertr::insist()}}
}
