% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flaw_report.R
\name{aud_report}
\alias{aud_report}
\title{Generate report of auditing results.}
\usage{
aud_report(.data)
}
\arguments{
\item{.data}{Dataset after being 'audited'.}
}
\value{
Returns a data frame of all failed audits.
}
\description{
Extracts relevant information from the \code{assertr_errors} attributes, after one
or more audits (\code{chk_*}) have been done.
}
\examples{

library(magrittr)
ds <- data.frame(a = rnorm(10), b = 1:10)
flaws <- ds \%>\%
chk_in_range(-1.1, 1.1, "a") \%>\%
chk_in_set(1:9, "b")
aud_report(flaws)

}
