% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicate_checks.R
\name{chk_duplicate}
\alias{chk_duplicate}
\title{Audits values in columns for duplicate values.}
\usage{
chk_duplicate(.data, .variables)
}
\arguments{
\item{.data}{Data to audit.}

\item{.variables}{Column variables to check.}
}
\value{
Adds attribute that contains the audit errors, if any, to the data.
Use \code{\link[=aud_report]{aud_report()}} to view the audit fails.
}
\description{
Among other things, uses \link[assertr:is_uniq]{assertr::is_uniq}.
}
\examples{
library(magrittr)

# dataset with one duplicate value between id and time
ds <- data.frame(
id = c(rep(1:3, each = 2), 3),
time = c(rep(1:2, times = 3), 1)
)
ds

flaws <- chk_duplicate(ds, c("id", "time"))
aud_report(flaws) # Should be one

flaws <- chk_duplicate(ds, c("id"))
aud_report(flaws) # Should be all of them

}
