% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_value_checks.R
\name{chk_in_range}
\alias{chk_in_range}
\title{Audits column values within a certain numeric range.}
\usage{
chk_in_range(.data, .low, .high, .variables)
}
\arguments{
\item{.data}{Data to audit.}

\item{.low}{Low value in the range.}

\item{.high}{High value in the range.}

\item{.variables}{Column variables to check.}
}
\value{
Adds attribute that contains the audit errors, if any, to the data.
Use \code{\link[=aud_report]{aud_report()}} to view the audit fails.
}
\description{
Uses the \code{\link[assertr:within_bounds]{assertr::within_bounds()}} function.
}
\examples{
library(magrittr)
ds <- data.frame(a = rnorm(10), b = rnorm(10))

flaws <- ds \%>\%
chk_in_range(-1, 1, "a")
aud_report(flaws)

}
