% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_value_checks.R
\name{chk_in_set}
\alias{chk_in_set}
\title{Audits column values of specific integers.}
\usage{
chk_in_set(.data, .values, .variables)
}
\arguments{
\item{.data}{Data to audit.}

\item{.values}{The integer values.}

\item{.variables}{Column variables to check.}
}
\value{
Adds attribute that contains the audit errors, if any, to the data.
Use \code{\link[=aud_report]{aud_report()}} to view the audit fails.
}
\description{
Uses the \code{\link[assertr:in_set]{assertr::in_set()}} function.
}
\examples{
library(magrittr)
ds <- data.frame(a = 1:3)

flaws <- ds \%>\%
chk_in_set(1:2, "a")
aud_report(flaws)
}
