% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_value_checks.R
\name{chk_missingness}
\alias{chk_missingness}
\title{Audit for any missingness in columns.}
\usage{
chk_missingness(.data, .variables)
}
\arguments{
\item{.data}{Data to audit.}

\item{.variables}{Column variables to check.}
}
\value{
Adds attribute that contains the audit errors, if any, to the data.
Use \code{\link[=aud_report]{aud_report()}} to view the audit fails.
}
\description{
Uses \code{\link[assertr:not_na]{assertr::not_na()}}.
}
\examples{

library(magrittr)
ds <- data.frame(a = c(1:4, NA, 6:10))
flaws <- ds \%>\%
chk_missingness("a")
aud_report(flaws)

}
