% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_value_checks.R
\name{chk_outliers}
\alias{chk_outliers}
\title{Audits for potential erroneous values ("outliers").}
\usage{
chk_outliers(.data, .n_mads, .variables)
}
\arguments{
\item{.data}{Data to audit.}

\item{.n_mads}{Number of mean absolute deviations (see \link[assertr:within_n_mads]{assertr::within_n_mads}).}

\item{.variables}{Column variables to check.}
}
\value{
Adds attribute that contains the audit errors, if any, to the data.
Use \code{\link[=aud_report]{aud_report()}} to view the audit fails.
}
\description{
Uses the \code{\link[assertr:within_n_mads]{assertr::within_n_mads()}}.
}
\examples{
library(magrittr)
ds <- data.frame(a = c(rnorm(100), 10))

flaws <- ds \%>\%
chk_outliers(2, "a")
aud_report(flaws)
}
