% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROMISE.audit-package.R
\docType{package}
\name{PROMISE.audit-package}
\alias{PROMISE.audit}
\alias{PROMISE.audit-package}
\title{PROMISE.audit: Identifies Unexpected Values in the PROMISE Datasets}
\description{
Checks given values of a dataset for "wrong" values (e.g. a DaysPerWeek
column should only contain integers 1 to 7 since there are 7 days in a week).
Adds these 'flaws' to the data so that a 'flaw' report can be created to look
into the violations in the raw data source.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lwjohnst86/PROMISE.audit}
  \item Report bugs at \url{https://github.com/lwjohnst86/PROMISE.audit/issues}
}

}
\author{
\strong{Maintainer}: Luke Johnston \email{lwjohnst@gmail.com}

}
\keyword{internal}
